test: 변경된 useInput, useInputs 에 맞춰 테스트코드 수정
