+++ hooks/src/lib/useCardHolder/useCardHolder.test.ts
@@ -3,9 +3,7 @@ import useCardHolder from "./useCardHolder";
 import { ERROR_MESSAGE } from "../constants";
 
 describe("신용카드의 소유자 이름 입력 테스트", () => {
-  const initialValue = {
-    name: "testName",
-  };
+  const initialValue = "testName";
 
   it("인자로 넘긴 값 그대로 초기값이 설정되어야 한다.", () => {
     const { result } = renderHook(() => useCardHolder(initialValue));
@@ -16,7 +14,7 @@ describe("신용카드의 소유자 이름 입력 테스트", () => {
   it("입력값 영어일 떄 입력한 값으로 업데이트 되어야 한다.", () => {
     const userInput = "changeName";
     const { result } = renderHook(() => useCardHolder(initialValue));
-    const target = { value: userInput, name: "name" };
+    const target = { value: userInput };
     const expectedValidationResult = {
       isValid: true,
       errorMessage: "",
@@ -29,14 +27,14 @@ describe("신용카드의 소유자 이름 입력 테스트", () => {
       } as React.ChangeEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.name).toEqual(userInput);
+    expect(result.current.inputValue).toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력값이 숫자일 때 업데이트가 안된다", () => {
     const userInput = "2353";
     const { result } = renderHook(() => useCardHolder(initialValue));
-    const target = { value: userInput, name: "name" };
+    const target = { value: userInput };
     const expectedValidationResult = {
       isValid: false,
       errorMessage: ERROR_MESSAGE.onlyEnglish,
@@ -49,14 +47,14 @@ describe("신용카드의 소유자 이름 입력 테스트", () => {
       } as React.ChangeEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.name).not.toEqual(userInput);
+    expect(result.current.inputValue).not.toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력값이 한글일 때 업데이트가 안된다", () => {
     const userInput = "한글 입력 테스트";
     const { result } = renderHook(() => useCardHolder(initialValue));
-    const target = { value: userInput, name: "name" };
+    const target = { value: userInput };
     const expectedValidationResult = {
       isValid: false,
       errorMessage: ERROR_MESSAGE.onlyEnglish,
@@ -69,14 +67,14 @@ describe("신용카드의 소유자 이름 입력 테스트", () => {
       } as React.ChangeEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.name).not.toEqual(userInput);
+    expect(result.current.inputValue).not.toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력 값이 모두 채워지고 블러 이벤트 발생 시 에러가 발생하지 않는다", () => {
     const userInput = "userInput";
     const { result } = renderHook(() => useCardHolder(initialValue));
-    const target = { value: userInput, name: "name" };
+    const target = { value: userInput };
     const expectedValidationResult = {
       isValid: true,
       errorMessage: "",
@@ -89,14 +87,14 @@ describe("신용카드의 소유자 이름 입력 테스트", () => {
       } as React.FocusEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.name).toEqual(userInput);
+    expect(result.current.inputValue).toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력 값이 없는 경우 블러 이벤트 발생 시 에러가 발생한다", () => {
     const userInput = "";
     const { result } = renderHook(() => useCardHolder(initialValue));
-    const target = { value: userInput, name: "name" };
+    const target = { value: userInput };
     const expectedValidationResult = {
       isValid: false,
       errorMessage: ERROR_MESSAGE.nameOutOfRange,
@@ -109,7 +107,7 @@ describe("신용카드의 소유자 이름 입력 테스트", () => {
       } as React.FocusEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.name).not.toEqual(userInput);
+    expect(result.current.inputValue).not.toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 });
