+++ hooks/src/lib/useCardNumber/useCardNumber.test.ts
@@ -4,9 +4,10 @@ import { ERROR_MESSAGE } from "../constants";
 
 describe("신용카드 번호 입력 테스트", () => {
   const initialValue = {
-    first: "first",
-    second: "second",
-    third: "third",
+    firstValue: "fi",
+    secondValue: "se",
+    thirdValue: "th",
+    fourthValue: "fo",
   };
 
   it("인자로 넘긴 값 그대로 초기값이 설정되어야 한다.", () => {
@@ -18,7 +19,7 @@ describe("신용카드 번호 입력 테스트", () => {
   it("입력값 숫자일 떄 입력한 값으로 업데이트 되어야 한다.", () => {
     const userInput = "1234";
     const { result } = renderHook(() => useCardNumber(initialValue));
-    const target = { value: userInput, name: "first" };
+    const target = { value: userInput, name: "firstValue" };
     const expectedValidationResult = {
       isValid: true,
       errorMessage: "",
@@ -31,14 +32,14 @@ describe("신용카드 번호 입력 테스트", () => {
       } as React.ChangeEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.first).toEqual(userInput);
+    expect(result.current.inputValue.firstValue).toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력값이 숫자가 아닐 때 업데이트가 안된다", () => {
     const userInput = "notUpdate";
     const { result } = renderHook(() => useCardNumber(initialValue));
-    const target = { value: userInput, name: "first" };
+    const target = { value: userInput, name: "firstValue" };
     const expectedValidationResult = {
       isValid: false,
       errorMessage: ERROR_MESSAGE.onlyNumber,
@@ -51,14 +52,14 @@ describe("신용카드 번호 입력 테스트", () => {
       } as React.ChangeEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.first).not.toEqual(userInput);
+    expect(result.current.inputValue.firstValue).not.toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력값이 4자리 이상일 때 업데이트가 안된다", () => {
     const userInput = "12345";
     const { result } = renderHook(() => useCardNumber(initialValue));
-    const target = { value: userInput, name: "first" };
+    const target = { value: userInput, name: "firstValue" };
     const expectedValidationResult = {
       isValid: false,
       errorMessage: ERROR_MESSAGE.onlyNumber,
@@ -71,14 +72,14 @@ describe("신용카드 번호 입력 테스트", () => {
       } as React.ChangeEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.first).not.toEqual(userInput);
+    expect(result.current.inputValue.firstValue).not.toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력 값이 모두 채워지고 블러 이벤트 발생 시 에러가 발생하지 않는다", () => {
     const userInput = "1234";
     const { result } = renderHook(() => useCardNumber(initialValue));
-    const target = { value: userInput, name: "first" };
+    const target = { value: userInput, name: "firstValue" };
     const expectedValidationResult = {
       isValid: true,
       errorMessage: "",
@@ -91,14 +92,14 @@ describe("신용카드 번호 입력 테스트", () => {
       } as React.FocusEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.first).toEqual(userInput);
+    expect(result.current.inputValue.firstValue).toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 
   it("입력 값이 모두 채워지지 않고 블러 이벤트 발생 시 에러가 발생한다", () => {
     const userInput = "125";
     const { result } = renderHook(() => useCardNumber(initialValue));
-    const target = { value: userInput, name: "first" };
+    const target = { value: userInput, name: "firstValue" };
     const expectedValidationResult = {
       isValid: false,
       errorMessage: ERROR_MESSAGE.cardNumberOutOfRange,
@@ -111,7 +112,7 @@ describe("신용카드 번호 입력 테스트", () => {
       } as React.FocusEvent<HTMLInputElement>);
     });
 
-    expect(result.current.inputValue.first).not.toEqual(userInput);
+    expect(result.current.inputValue.firstValue).not.toEqual(userInput);
     expect(result.current.validationResult).toEqual(expectedValidationResult);
   });
 });
